requirements.txt
python-telegram-bot==20.3
telethon
python-dotenv
requests
psutil
supervisor
